# [aiohttp] - use batch updates