@Reskov 2025-06-06
also `[]` is slightly faster than `tuple`
```
tuple(zip(update_ids, numbers)): 0.764148 seconds
[*(zip(update_ids, numbers))]: 0.721439 seconds
Difference: 0.042709 seconds
Method 2 is faster by 5.92%
```


```
from random import sample, randint
import timeit

# Define number of queries
num_queries = 20

update_ids = sample(range(1, 10001), num_queries)
update_ids.sort()

numbers = [randint(1, 10000) for _ in range(num_queries)]
fetch_params = [(i,) for i in update_ids]

# Benchmark the two approaches
def method1():
    return tuple(zip(update_ids, numbers))

def method2():
    return [*(zip(update_ids, numbers))]

# Run benchmark
iterations = 1000000
time1 = timeit.timeit(method1, number=iterations)
time2 = timeit.timeit(method2, number=iterations)

print(f"tuple(zip(update_ids, numbers)): {time1:.6f} seconds")
print(f"[*(zip(update_ids, numbers))]: {time2:.6f} seconds")
print(f"Difference: {abs(time1-time2):.6f} seconds")
print(f"Method {'1' if time1 < time2 else '2'} is faster by {(max(time1,time2)/min(time1,time2)-1)*100:.2f}%")
```